sync RMF to 0fdafa08866d0f5f1378d33354546deaae892a03
git-svn-id: https://svn.salilab.org/imp/trunk@16766 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/rmf/dependency/RMF/src/FrameHandle.cpp
@@ -26,7 +26,7 @@ FrameHandle FrameHandle::add_child(std::string name, FrameType t) {
                   get_shared_data());
   ret.set_as_current_frame();
   RMF_INTERNAL_CHECK(get_shared_data()->get_number_of_frames()
-                     == ret.get_id().get_index() + 1,
+                     == static_cast<unsigned int>(ret.get_id().get_index() + 1),
                      "Wrong number of frames");
   return ret;
 }
